+++ NEWS.md
@@ -1,4 +1,5 @@
-* Added the functin `parse_ks_sql`
+* Added the function `cat2int`
+* Added the function `parse_ks_sql`
 
 # add2impala 0.0.0.9000